:sparkles: MyTradeHistoryListContent, List 컴포넌트 분리
