+++ src/app/(root)/(routes)/mypage/histories/components/my-trade-history-list-content/index.tsx
@@ -1,0 +1,3 @@
+import MyTradeHistoryListContent from './MyTradeHistoryListContent'
+
+export default MyTradeHistoryListContent
