+++ src/app/(root)/(routes)/mypage/histories/components/my-trade-history-list/index.tsx
@@ -1,3 +1,3 @@
-import HistoryList from './MyTradeHistoryList'
+import MyTradeHistoryList from './MyTradeHistoryList'
 
-export default HistoryList
+export default MyTradeHistoryList
